Move types to constants
