+++ client/src/containers/projectDetail.ts
@@ -2,7 +2,7 @@ import { connect, Dispatch } from "react-redux";
 import {withRouter} from "react-router-dom";
 import {includes} from 'lodash'
 
-import { AppState } from "../types/index";
+import { AppState } from "../constants/types";
 import ProjectDetail from "../components/projectDetail";
 import * as actions from "../actions/project";
