+++ client/src/reducers/app.ts
@@ -1,7 +1,7 @@
 import {combineReducers} from "redux";
 import {projectsReducer} from "./projects";
 import {experimentsReducer} from "./experiments";
-import {AppState} from "../types/index";
+import {AppState} from "../constants/types";
 
 
 const appReducer = combineReducers<AppState>({
